// src/core/modules/dao-api/models/CreateDeviceRequest.ts
/**
 * DAO-API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Input information needed to create a new API key
*/
export class CreateDeviceRequest {
    /**
    * Name for the API key
    */
    'deviceName'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "deviceName",
            "baseName": "device_name",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateDeviceRequest.attributeTypeMap;
    }

    public constructor() {
    }
}